requirements.txt
streamlit>=1.30
requests>=2.31
pandas>=2.2
python-dotenv>=1.0
matplotlib>=3.8
seaborn>=0.13
plotly>=6.0
nbformat>=5.10
ipywidgets>=8.1
